+++ tests/features/test_feat.py
@@ -14,9 +14,8 @@ from ididi import DependencyGraph, entry
 dg = DependencyGraph()
 
 
-class DataBase:
-    def __init__(self, url: str = ""):
-        self.url = url
+class ResourceBase:
+    def __init__(self):
         self._is_closed = False
 
     async def close(self) -> None:
@@ -27,9 +26,43 @@ class DataBase:
         return self._is_closed
 
 
+class AsyncResourceBase:
+    def __init__(self):
+        self._is_closed = False
+
+    async def close(self) -> None:
+        self._is_closed = True
+
+    @property
+    def is_closed(self) -> bool:
+        return self._is_closed
+
+
+class Client(AsyncResourceBase):
+    def __init__(self, client_type: str = "postgres"):
+        super().__init__()
+        self.client_type = client_type
+
+
+class DataBase(AsyncResourceBase):
+    def __init__(self, client: Client):
+        super().__init__()
+        self.client = client
+
+
 @dg.node
-async def get_db() -> ty.AsyncGenerator[DataBase, None]:
-    db = DataBase(url="sqlite://")
+async def get_client() -> ty.AsyncGenerator[Client, None]:
+    client = Client()
+    try:
+        yield client
+    finally:
+        await client.close()
+
+
+@dg.node
+async def get_db(client: Client) -> ty.AsyncGenerator[DataBase, None]:
+    db = DataBase(client)
+    assert not client.is_closed
     try:
         yield db
     finally:
@@ -45,31 +78,3 @@ async def main(db: DataBase):
 @pytest.mark.asyncio
 async def test_async_gen_factory():
     await main()
-    # db = dg.resolve(get_db)
-    # breakpoint()
-
-    # dg = DependencyGraph()
-
-    # class DataBase:
-    #     def __init__(self, url: str = ""):
-    #         self.url = url
-    #         self._is_closed = False
-
-    #     async def close(self) -> None:
-    #         self._is_closed = True
-
-    #     @property
-    #     def is_closed(self) -> bool:
-    #         return self._is_closed
-
-    # @dg.node(reuse=False)
-    # @asynccontextmanager
-    # async def get_db() -> ty.AsyncGenerator[DataBase, None]:
-    #     db = DataBase(url="sqlite://")
-    #     try:
-    #         yield db
-    #     finally:
-    #         await db.close()
-
-    # async def main():
-    #     dg.static_resolve(get_db)